requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.11.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
phonenumberslite==8.13.49
psycopg2-binary==2.9.10
PyJWT==2.10.0
sqlparse==0.5.1
